requirements.txt
Flask~=3.0.3
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1
Flask-SQLAlchemy
pymysql